+++ Changelog.md
@@ -1,5 +1,12 @@
-﻿Changelog
+Changelog
 =========
+
+# Version 1
+- Name of program is now "Scammer Bingo Unix"
+
+
+# Alpha 4
+- Name changes for title
 
 # Alpha 3:
 - Made the UI look a little better, added a background and stuff